moves login/register to a component
